Consolidate participant creation trigger logic into single trigger function
