+++ functions/src/participant.utils.ts
@@ -6,15 +6,21 @@ import {
   ParticipantStatus,
   TransferStageConfig,
   StageConfig,
+  StageKind,
   CohortConfig,
   createCohortConfig,
   createMetadataConfig,
   SurveyStagePublicData,
   SurveyQuestionKind,
   MultipleChoiceSurveyAnswer,
+  createChipStageParticipantAnswer,
+  createPayoutStageParticipantAnswer,
 } from '@deliberation-lab/utils';
 import {completeStageAsAgentParticipant} from './agent_participant.utils';
-import {getFirestoreActiveParticipants} from './utils/firestore';
+import {
+  getFirestoreActiveParticipants,
+  getFirestoreParticipant,
+} from './utils/firestore';
 import {generateId} from '@deliberation-lab/utils';
 import {createCohortInternal} from './cohort.utils';
 
@@ -394,4 +400,92 @@ export async function getParticipantRecord(
   return participantSnapshot.exists
     ? (participantSnapshot.data() as ParticipantProfileExtended)
     : null;
+}
+
+/** Set up participant stage answers. */
+export async function initializeParticipantStageAnswers(
+  experimentId: string,
+  participantConfig: ParticipantProfileExtended,
+) {
+  await app.firestore().runTransaction(async (transaction) => {
+    const participantId = participantConfig.privateId;
+
+    // Get all stage configs
+    const stageConfigs = (
+      await app
+        .firestore()
+        .collection(`experiments/${experimentId}/stages`)
+        .get()
+    ).docs.map((doc) => doc.data() as StageConfig);
+
+    const getRandomChipValue = (chip: ChipItem) => {
+      const step = 0.1;
+      const lower = Math.round(chip.lowerValue / step);
+      const upper = Math.round(chip.upperValue / step);
+      const randomStep =
+        Math.floor(Math.random() * (upper - lower + 1)) + lower;
+
+      return parseFloat((randomStep * step).toFixed(2));
+    };
+
+    for (const stage of stageConfigs) {
+      // Define doc reference for stage
+      const stageDoc = app
+        .firestore()
+        .collection('experiments')
+        .doc(experimentId)
+        .collection('participants')
+        .doc(participantId)
+        .collection('stageData')
+        .doc(stage.id);
+
+      // Write stage answer if relevant
+      switch (stage.kind) {
+        case StageKind.CHIP:
+          // If chip stage, set default chips for participant based on config
+          const chipMap: Record<string, number> = {};
+          const chipValueMap: Record<string, number> = {};
+          stage.chips.forEach((chip) => {
+            chipMap[chip.id] = chip.startingQuantity;
+            chipValueMap[chip.id] = getRandomChipValue(chip);
+          });
+
+          const chipAnswer = createChipStageParticipantAnswer(
+            stage.id,
+            chipMap,
+            chipValueMap,
+          );
+
+          transaction.set(stageDoc, chipAnswer);
+
+          // Set public stage data
+          const publicChipDoc = app
+            .firestore()
+            .collection('experiments')
+            .doc(experimentId)
+            .collection('cohorts')
+            .doc(participantConfig.currentCohortId)
+            .collection('publicStageData')
+            .doc(stage.id);
+
+          const publicChipData = (
+            await publicChipDoc.get()
+          ).data() as ChipStagePublicData;
+          const publicId = participantConfig.publicId;
+
+          publicChipData.participantChipMap[publicId] = chipAnswer.chipMap;
+          publicChipData.participantChipValueMap[publicId] =
+            chipAnswer.chipValueMap;
+          transaction.set(publicChipDoc, publicChipData);
+          break;
+        case StageKind.PAYOUT:
+          // If payout stage, set random selection of payout items
+          const payoutAnswer = createPayoutStageParticipantAnswer(stage);
+          transaction.set(stageDoc, payoutAnswer);
+          break;
+        default:
+          break;
+      }
+    } // end stage config logic
+  }); // end transaction
 }
